requirements.txt
Django
djangorestframework==3.14.0
djangorestframework-simplejwt
dj-database-url==2.1.0
psycopg2-binary==2.9.10
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.0
celery==5.3.4
redis
django-redis
boto3==1.29.7
django-storages==1.14.2
pillow==11.2.1
django-filter==23.3
drf-spectacular==0.26.5
python-decouple==3.8
sqlparse==0.5.3
django-environ==0.11.2
django-storage-supabase==0.1.0 
setuptools
drf-yasg
